# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TaskFlow API",
        "description": "API documentation for TaskFlow project",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local server"
        }
    ],
    "paths": {
        "/api/menus": {
            "get": {
                "tags": [
                    "Menu"
                ],
                "summary": "Get all menus",
                "operationId": "1dc6b7b84ebea40905c68ce4f844da25",
                "responses": {
                    "200": {
                        "description": "List of all menus",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Menu"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Menu"
                ],
                "summary": "Create a new menu",
                "operationId": "37652f848e93b047d0f6f39cb4df6269",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Dashboard"
                                    },
                                    "icon": {
                                        "type": "string",
                                        "example": "home"
                                    },
                                    "url": {
                                        "type": "string",
                                        "example": "/dashboard"
                                    },
                                    "order": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "is_active": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Menu successfully created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Menu"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/menus/{id}": {
            "get": {
                "tags": [
                    "Menu"
                ],
                "summary": "Get a specific menu by ID",
                "operationId": "358e9ead3d3f6d8fd5b51d1f0882dc83",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Menu ID (UUID)",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Menu found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Menu"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Menu not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Menu"
                ],
                "summary": "Update an existing menu",
                "operationId": "67ceffba16f3b7be4391a4d6681fa431",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Menu ID (UUID)",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Updated Dashboard"
                                    },
                                    "icon": {
                                        "type": "string",
                                        "example": "dashboard-outline"
                                    },
                                    "url": {
                                        "type": "string",
                                        "example": "/dashboard"
                                    },
                                    "order": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "is_active": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Menu successfully updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Menu"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Menu not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Menu"
                ],
                "summary": "Delete a menu",
                "operationId": "abff8a90938116c08c448319c3dc7323",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Menu ID (UUID)",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Menu successfully deleted"
                    },
                    "404": {
                        "description": "Menu not found"
                    }
                }
            }
        },
        "/api/menu-profile/attach": {
            "post": {
                "tags": [
                    "MenuProfile"
                ],
                "summary": "Attach a Menu to a Profile",
                "description": "Associates a specific menu item to a profile using their UUIDs.",
                "operationId": "706f05b5810a725755143a59e8a58490",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "profile_id",
                                    "menu_id"
                                ],
                                "properties": {
                                    "profile_id": {
                                        "type": "string",
                                        "format": "uuid",
                                        "example": "4da9934b-cede-4f05-ba1c-38ab1f55ccf9"
                                    },
                                    "menu_id": {
                                        "type": "string",
                                        "format": "uuid",
                                        "example": "9da9934b-cede-4f05-ba1c-38ab1f55ccf9"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Menu successfully attached to the profile."
                    },
                    "422": {
                        "description": "Validation error."
                    }
                }
            }
        },
        "/api/menu-profile/detach": {
            "post": {
                "tags": [
                    "MenuProfile"
                ],
                "summary": "Detach a Menu from a Profile",
                "description": "Removes the association between a menu and a profile.",
                "operationId": "39ea4b5a1656b5e6b19c13e412e6635b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "profile_id",
                                    "menu_id"
                                ],
                                "properties": {
                                    "profile_id": {
                                        "type": "string",
                                        "format": "uuid",
                                        "example": "4da9934b-cede-4f05-ba1c-38ab1f55ccf9"
                                    },
                                    "menu_id": {
                                        "type": "string",
                                        "format": "uuid",
                                        "example": "9da9934b-cede-4f05-ba1c-38ab1f55ccf9"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Menu successfully detached from the profile."
                    },
                    "422": {
                        "description": "Validation error."
                    }
                }
            }
        },
        "/api/menu-profile/{profile_id}": {
            "get": {
                "tags": [
                    "MenuProfile"
                ],
                "summary": "Get Menus by Profile",
                "description": "Retrieves all menus associated with a specific profile.",
                "operationId": "a5dba6dd9aa2d29731a55c8b389e76fa",
                "parameters": [
                    {
                        "name": "profile_id",
                        "in": "path",
                        "description": "UUID of the profile",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of menus retrieved successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Menu"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Profile not found."
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Menu": {
                "title": "Menu",
                "description": "Represents a menu item within the system.",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "example": "9da9934b-cede-4f05-ba1c-38ab1f55ccf9"
                    },
                    "name": {
                        "type": "string",
                        "example": "Dashboard"
                    },
                    "slug": {
                        "type": "string",
                        "example": "dashboard"
                    },
                    "icon": {
                        "type": "string",
                        "example": "home"
                    },
                    "url": {
                        "type": "string",
                        "example": "/dashboard"
                    },
                    "order": {
                        "type": "integer",
                        "example": 1
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-11-03T12:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-11-03T12:10:00Z"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}